Refactor slightly
