+++ app/helpers/react_components/mentoring/session.rb
@@ -32,7 +32,7 @@ module ReactComponents
             ),
             mentor_solution: mentor_solution,
             exemplar_solution: exercise.exemplar_files.values.first,
-            notes: exercise.mentoring_notes.to_s,
+            notes: exercise.mentoring_notes.content,
             out_of_date: solution.out_of_date?,
             download_command: solution.mentor_download_cmd,
             scratchpad: {
@@ -60,7 +60,7 @@ module ReactComponents
         {
           mentor_dashboard: Exercism::Routes.mentoring_inbox_path,
           exercise: Exercism::Routes.track_exercise_path(track, exercise),
-          improve_notes: exercise.edit_mentoring_notes_url,
+          improve_notes: exercise.mentoring_notes.edit_url,
           mentoring_docs: Exercism::Routes.docs_section_path(:mentoring)
         }
       end
