+++ test/helpers/react_components/mentoring/session_test.rb
@@ -57,7 +57,7 @@ module Mentoring
           links: {
             mentor_dashboard: Exercism::Routes.mentoring_inbox_path,
             exercise: Exercism::Routes.track_exercise_path(track, exercise),
-            improve_notes: exercise.edit_mentoring_notes_url,
+            improve_notes: exercise.mentoring_notes.edit_url,
             mentoring_docs: Exercism::Routes.docs_section_path(:mentoring)
           }
         }
@@ -124,7 +124,7 @@ module Mentoring
           links: {
             mentor_dashboard: Exercism::Routes.mentoring_inbox_path,
             exercise: Exercism::Routes.track_exercise_path(track, exercise),
-            improve_notes: exercise.edit_mentoring_notes_url,
+            improve_notes: exercise.mentoring_notes.edit_url,
             mentoring_docs: Exercism::Routes.docs_section_path(:mentoring)
           }
         }
@@ -188,7 +188,7 @@ module Mentoring
           links: {
             mentor_dashboard: Exercism::Routes.mentoring_inbox_path,
             exercise: Exercism::Routes.track_exercise_path(track, exercise),
-            improve_notes: exercise.edit_mentoring_notes_url,
+            improve_notes: exercise.mentoring_notes.edit_url,
             mentoring_docs: Exercism::Routes.docs_section_path(:mentoring)
           }
         }
