+++ test/models/git/exercise/mentor_notes_test.rb
@@ -1,12 +1,13 @@
 require "test_helper"
 
-class Exercise::MentorNotesTest < ActiveSupport::TestCase
+class Git::Exercise::MentorNotesTest < ActiveSupport::TestCase
   test "#edit_url returns edit url if mentoring notes is present" do
     track = create :track
     exercise = create :concept_exercise, track: track
-    repo = mock
+    repo = TestHelpers.use_website_copy_test_repo!
     repo.expects(:mentor_notes_for).with(track.slug, exercise.slug).returns("mentor notes")
-    notes = Exercise::MentorNotes.new(exercise, copy_repo: repo)
+
+    notes = Git::Exercise::MentorNotes.new(track.slug, exercise.slug)
 
     assert_equal(
       "https://github.com/exercism/website-copy/edit/main/tracks/ruby/exercises/strings/mentoring.md",
@@ -17,9 +18,10 @@ class Exercise::MentorNotesTest < ActiveSupport::TestCase
   test "#edit_url returns new url if mentoring notes is blank" do
     track = create :track
     exercise = create :concept_exercise, track: track
-    repo = mock
+    repo = TestHelpers.use_website_copy_test_repo!
     repo.expects(:mentor_notes_for).with(track.slug, exercise.slug).returns("")
-    notes = Exercise::MentorNotes.new(exercise, copy_repo: repo)
+
+    notes = Git::Exercise::MentorNotes.new(track.slug, exercise.slug)
 
     assert_equal(
       "https://github.com/exercism/website-copy/new/main/tracks/ruby/exercises/strings?filename=strings/mentoring.md",
