+++ test/test_helper.rb
@@ -60,8 +60,9 @@ module TestHelpers
 
   def self.use_website_copy_test_repo!
     repo_url = TestHelpers.git_repo_url("website-copy")
-    repo = Git::WebsiteCopy.new(repo_url: repo_url)
-    Git::WebsiteCopy.expects(:new).at_least_once.returns(repo)
+    Git::WebsiteCopy.new(repo_url: repo_url).tap do |repo|
+      Git::WebsiteCopy.expects(:new).at_least_once.returns(repo)
+    end
   end
 
   def self.use_blog_test_repo!
